# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.GenerateUploadURL:
    properties:
      content_type:
        enum:
        - image/jpeg
        - image/png
        - image/webp
        type: string
      expiry_seconds:
        maximum: 86400
        minimum: 1
        type: integer
      file_name:
        maxLength: 255
        minLength: 1
        type: string
      file_size:
        maximum: 10485760
        minimum: 1
        type: integer
      resource_id:
        minimum: 0
        type: integer
      upload_type:
        enum:
        - avatar
        - post_cover
        - post_content
        type: string
    required:
    - content_type
    - file_name
    - file_size
    - upload_type
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.SendNotification:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
      user_id:
        minimum: 1
        type: integer
    required:
    - content
    - title
    - user_id
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileDetail:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      object_key:
        type: string
      resource_id:
        type: integer
      uploader_id:
        type: integer
      url:
        type: string
      usage:
        type: string
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileUploadURL:
    properties:
      expires:
        type: integer
      file_id:
        type: integer
      object_key:
        type: string
      upload_url:
        type: string
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: Xiaoqing Blog 后端 API（Admin + Public V1）。
  title: Xiaoqing Blog API
  version: "1.0"
paths:
...
  /admin/files/:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        enum:
        - created_at
        - file_size
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 用途过滤
        in: query
        name: filter.usage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 文件列表
      tags:
      - Admin.Files
  /admin/files/{object_key}:
    delete:
      parameters:
      - description: 对象 Key
        in: path
        name: object_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 删除文件
      tags:
      - Admin.Files
  /admin/files/upload-url:
    post:
      consumes:
      - application/json
      parameters:
      - description: 上传参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.GenerateUploadURL'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FileUploadURL'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 生成文件上传 URL
      tags:
      - Admin.Files
...
  /admin/notifications/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 通知内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.SendNotification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 发送通知
      tags:
      - Admin.Notifications
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"